refactor(cli): move generate's option groups out of the capped file
cli/generate.py sat at 995 lines against a 1000-line hard gate, so the next
flag added to it would have failed `make file-length` for a reason unrelated
to its own change. Nearly all of it was the ~50-option decorator stack and a
block of flag-resolution helpers, neither of which has to live next to the
command body.

The option stack moves to cli/generate_options.py as six group decorators —
scope, output, run, selection, per-memory-type, automation — each a contiguous
slice of the stack in its original order, because Click builds `--help` and
`make docs-cli` builds the CLI reference from the order options were applied.
The helpers move to cli/generate_resolution.py, the layer above _date_resolution
that decides which flags outrank the config file and which combinations
contradict each other.

register_generate_commands is untouched: its 163 cognitive-complexity points
are a separate problem from its length, and mixing the two would hide both.
The only lines added to generate.py are its import block and the six decorator
lines; everything else is a deletion of a moved block.

Proven behaviour-neutral by dumping `main.commands["generate"].params` — name,
opts, type, default, help, hidden, flag value, multiple, callback, order — on
both sides of the change and diffing: identical but for two heap addresses in
click.types.Path's repr, whose attributes were compared separately and match.

595 / 351 / 168 lines.

Refs #686
